requirements.txt
opencv-python
customtkinter
pillow
pyserial
pymongo
python-dotenv
google-generativeai
openai
pydantic
ultralytics
